Fix: Workbook API 오류 수정 (#198)
* Fix: switch문에 case별 break 추가 #197

* Feat: MYSQL Keyword와 동일한 명칭으로 인해 오류 발생, 칼럼명 변환 및 관련 API 적용 #197

* Style: 주석 오탈자 수정 #197

* Fix: Entity 칼럼명 변경으로 인한 쿼리 변경 #197

* Fix: 페이지네이션 API 오류 수정 #197

* Feat: 최신순 기반 페이지네이션 적용 #197

+++ src/main/java/com/selfrunner/gwalit/domain/workbook/service/WorkbookService.java
@@ -3,23 +3,22 @@ package com.selfrunner.gwalit.domain.workbook.service;
 import com.selfrunner.gwalit.domain.member.entity.Member;
 import com.selfrunner.gwalit.domain.workbook.dto.request.WorkbookReq;
 import com.selfrunner.gwalit.domain.workbook.dto.response.*;
+import com.selfrunner.gwalit.domain.workbook.entity.SubjectDetail;
 import com.selfrunner.gwalit.domain.workbook.entity.Views;
 import com.selfrunner.gwalit.domain.workbook.entity.Workbook;
 import com.selfrunner.gwalit.domain.workbook.exception.WorkbookException;
 import com.selfrunner.gwalit.domain.workbook.repository.ViewsRepository;
 import com.selfrunner.gwalit.domain.workbook.repository.WorkbookRepository;
-import com.selfrunner.gwalit.global.exception.ApplicationException;
 import com.selfrunner.gwalit.global.exception.ErrorCode;
 import com.selfrunner.gwalit.global.util.aws.S3Client;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -87,24 +86,23 @@ public class WorkbookService {
         Workbook workbook = workbookRepository.findById(workbookId).orElseThrow(() -> new WorkbookException(ErrorCode.NOT_FOUND_EXCEPTION));
 
         // Business Logic: 이미지, 파일 등 등록 이후, Entity로 업데이트
-        String thumbnailUrl, workbookFileUrl, answerFileUrl;
         // 문제집 파일을 직접 올릴 경우
         if(workbookReq.getIsFile().equals(Boolean.TRUE)) {
             try {
-                if(!thumbnailImage.isEmpty()) {
+                if(thumbnailImage != null && !thumbnailImage.isEmpty()) {
                     s3Client.delete(workbook.getThumbnailUrl());
                     String updateThumbnailImageUrl = s3Client.upload(thumbnailImage, "workbook/" + workbookReq.getType());
                     workbook.updateThumbnailUrl(updateThumbnailImageUrl);
                 }
-                if(!workbookFile.isEmpty()) {
+                if(workbookFile != null && !workbookFile.isEmpty()) {
                     s3Client.delete(workbook.getWorkbookFileUrl());
                     String updateWorkbookFileUrl = s3Client.upload(workbookFile, "workbook/" + workbookReq.getType());
-                    workbook.updateThumbnailUrl(updateWorkbookFileUrl);
+                    workbook.updateWorkbookFileUrl(updateWorkbookFileUrl);
                 }
-                if(!answerFile.isEmpty()) {
+                if(answerFile != null && !answerFile.isEmpty()) {
                     s3Client.delete(workbook.getAnswerFileUrl());
                     String updateAnswerFileUrl = s3Client.upload(answerFile, "workbook/" + workbookReq.getType());
-                    workbook.updateThumbnailUrl(updateAnswerFileUrl);
+                    workbook.updateAnswerFileUrl(updateAnswerFileUrl);
                 }
                 workbook.update(workbookReq);
 
@@ -118,7 +116,7 @@ public class WorkbookService {
         // 문제를 별도로 등록하여, 참조 테이블로 엮어서 만드는 경우
         if(workbookReq.getIsFile().equals(Boolean.FALSE)) {
             try {
-                if(!thumbnailImage.isEmpty()) {
+                if(thumbnailImage != null && !thumbnailImage.isEmpty()) {
                     s3Client.delete(workbook.getThumbnailUrl());
                     String updateThumbnailImageUrl = s3Client.upload(thumbnailImage, "workbook/" + workbookReq.getType());
                     workbook.updateThumbnailUrl(updateThumbnailImageUrl);
@@ -190,13 +188,33 @@ public class WorkbookService {
         return workbookCardResList;
     }
 
-    public Slice<WorkbookCardRes> getWorkbookList(Member member, String subjectDetail, String type, Long cursor, Pageable pageable) {
+    public Slice<WorkbookCardRes> getWorkbookList(Member member, String detail, String type, Long cursor, Pageable pageable) {
         // Validation: Controller에서 @Auth를 통해 인증된 사용자만 접근할 수 있도록 함.
 
         // Business Logic
-        Slice<WorkbookCardRes> workbookCardResSlice = workbookRepository.findWorkbookCardPageableBy(subjectDetail, type, cursor, pageable);
+        SubjectDetail subjectDetail = getSubjectDetail(detail);
+
+        LocalDateTime cursorCreatedAt = (cursor != null) ? workbookRepository.findById(cursor).get().getCreatedAt() : null;
+        
+        Slice<WorkbookCardRes> workbookCardResSlice = workbookRepository.findWorkbookCardPageableBy(subjectDetail, type, cursor, cursorCreatedAt, pageable);
 
         // Response
         return workbookCardResSlice;
+    }
+
+    // detail 과목 정보 조회 메소드
+    private static SubjectDetail getSubjectDetail(String detail) {
+        SubjectDetail subjectDetail = null;
+        SubjectDetail[] subjectDetailList = SubjectDetail.values();
+        for(SubjectDetail temp : subjectDetailList) {
+            if(temp.getEnglishName().equals(detail)) {
+                subjectDetail = temp;
+                break;
+            }
+        }
+        if(subjectDetail == null) {
+            throw new WorkbookException(ErrorCode.INVALID_VALUE_EXCEPTION);
+        }
+        return subjectDetail;
     }
 }
